Fix misleading documentation & some typos in the readme
onSingleTap and onLongTap is documented as "Called when ... click on
marker" while it calls the function when the user clicks on the map
and not on a marker. If the click happens exactly on the marker, the
function is not called. That's why I would like to change the
documentation to "Called when ... click on map".

This commit also fixes some typos in the readme.